services/notifications_service/requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
celery==5.2.3
django==3.2.5
djangorestframework==3.12.4
twilio==6.59.1
django-celery-beat==2.2.1
django-celery-results==2.4.0
